refactor: move UpdateNicknameModal component
