+++ front-end/src/renderer/components/modals/UpdateNicknameModal.vue
